% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_completeness.R
\name{distant_neg_val}
\alias{distant_neg_val}
\title{Find highly distant value for data frame}
\usage{
distant_neg_val(data)
}
\arguments{
\item{data}{data frame.}
}
\value{
Numeric vector of length 1
}
\description{
Returns a numeric value which is distant from the values in \code{data} using
the following equation: \deqn{output = -2 * (max(data)-min(data)))}
}
\examples{
data(mtcars)
eHDPrep:::distant_neg_val(mtcars)
}
