% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_clean.R
\name{encode_genotype_vec}
\alias{encode_genotype_vec}
\title{Encode a genotype/SNP vector}
\usage{
encode_genotype_vec(x)
}
\arguments{
\item{x}{input vector containing genotype data}
}
\value{
Ordered factor, ordered by allele frequency in variable
}
\description{
Standardises homozygous SNP alleles (e.g. recorded as 'A') to two character
form (e.g. 'A/A') and orders heterozygous SNP alleles alphabetically (e.g.
"GA" becomes "A/G"). The SNP values are then converted from a character
vector to an ordered factor, ordered by SNP allele frequency (e.g. most
frequent SNP allele is 1, second most frequent value is 2, and least frequent
values is 3). This method embeds the numeric relationship between the SNP
allele frequencies while preserving value labels.
}
\examples{
data(example_data)
vec <- example_data[,"SNP_a",drop = TRUE]
eHDPrep:::encode_genotype_vec(vec)
}
