% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_edge_tbl}
\alias{example_edge_tbl}
\title{Example ontology as an edge table for semantic enrichment}
\format{
A data frame: \describe{
\item{from}{character. Names of semantic concepts which have a directed relationship to concepts in `to` column.}
\item{to}{character. Names of semantic concepts which have a directed relationship to concepts in `from` column.}
}
}
\source{
synthetic
}
\usage{
example_edge_tbl
}
\description{
A data frame describing semantic links (edges) between entities in
`example_ontology`. Used to demonstrate semantic enrichment.
}
\details{
Used in documentation and creation of `example_ontology` in 'eHDPrep'.
}
\keyword{datasets}
