% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ontology}
\alias{example_ontology}
\title{Example ontology as a network graph for semantic enrichment}
\format{
tidygraph graph
}
\source{
synthetic
}
\usage{
example_ontology
}
\description{
A small custom network graph to demonstrate semantic enrichment.
}
\details{
Contains semantic links of variables in eHDPrep's `example_data` following
quality control.
}
\keyword{datasets}
