% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_info_theory.R
\name{information_content_contin}
\alias{information_content_contin}
\title{Calculate Information Content (Continuous Variable)}
\usage{
information_content_contin(x, unit = c("bits"))
}
\arguments{
\item{x}{Input vector}

\item{unit}{Unit to measure entropy. Either "bits" (default) of "nats".}
}
\value{
Information content of input variable
}
\description{
Calculates information content of a continuous (numeric) vector in bits
(default) or natural units. Missing values are omitted from the calculation.
}
\examples{
data(example_data)
information_content_contin(example_data$tumoursize)
}
