% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_info_theory.R
\name{information_content_discrete}
\alias{information_content_discrete}
\title{Calculate Information Content (Discrete Variable)}
\usage{
information_content_discrete(x, unit = c("bits"))
}
\arguments{
\item{x}{Input vector}

\item{unit}{Unit to measure entropy. Either "bits" (default) or "nats".}
}
\value{
Information content of input variable
}
\description{
Calculates information content of a discrete (categorical or ordinal) vector
in bits (default) or natural units. Missing values are omitted from the
calculation.
}
\examples{
data(example_data)
information_content_discrete(example_data$marital_status)
}
