% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_enrich.R
\name{max_catchNAs}
\alias{max_catchNAs}
\title{Find maximum of vector safely}
\usage{
max_catchNAs(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
maximum value of \code{x}
}
\description{
This low-level function is deployed as part of the semantic enrichment
process.Calculates maximum of values in numeric vector (ignoring NAs). If all
values in input vector are \code{NA}, returns \code{NA} (rather than -Inf),
}
\examples{
x <- c(1,2,3)
eHDPrep:::max_catchNAs(x)
}
