% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_info_theory.R
\name{mi_content_discrete}
\alias{mi_content_discrete}
\title{Calculate Mutual Information Content}
\usage{
mi_content_discrete(x, y)
}
\arguments{
\item{x}{First variable}

\item{y}{Second variable}
}
\value{
Mutual information content of \code{x} and \code{y}
}
\description{
Calculates mutual information content between two variables in bits. Missing
values are omitted from the calculation.
}
\examples{
data(example_data)
mi_content_discrete(example_data$diabetes, example_data$diabetes_type)
}
