% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_enrich.R
\name{min_catchNAs}
\alias{min_catchNAs}
\title{Find minimum of vector safely}
\usage{
min_catchNAs(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
minimum value of \code{x}
}
\description{
This low-level function is deployed as part of the semantic enrichment
process. Calculates minimum of values in numeric vector (ignoring NAs). If
all values in numeric vector are \code{NA}, returns \code{NA} (rather than
Inf),
}
