% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_enrich.R
\name{sum_catchNAs}
\alias{sum_catchNAs}
\title{Sum vector safely for semantic enrichment}
\usage{
sum_catchNAs(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sum of \code{x}
}
\description{
sums values in x (ignoring NAs). If all values in x are \code{NA}, returns
\code{NA} (rather than 0),
}
\examples{
x <- c(1,2,3)
eHDPrep:::sum_catchNAs(x)
}
