% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_enrich.R
\name{validate_mapping_tbl}
\alias{validate_mapping_tbl}
\title{Validate mapping table for semantic enrichment}
\usage{
validate_mapping_tbl(mapping_tbl, data, ontol_graph)
}
\arguments{
\item{mapping_tbl}{data frame. Contains two columns. First column contains
variable names of a primary dataset. Second column contains entities in
an ontological graph to which the primary dataset's variable names are mapped.}

\item{data}{data frame. Primary dataset which contains variable names
referred to in first column of the mapping table}

\item{ontol_graph}{ontological graph which contains entity names/IDs referred
to in second column of the mapping table}
}
\value{
Any warnings and the mapping table returned invisibly
}
\description{
Applies tests to a mapping table to ensure it is valid for use with
the data frame and ontological graph, in preparation for semantic enrichment.
}
\examples{
data(example_mapping_file)
data(example_data)
data(example_ontology)

# warns that some variables don't correspond between the data and the mapping file:
eHDPrep:::validate_mapping_tbl(example_mapping_file, example_data, example_ontology)
}
