% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semantic_enrich.R
\name{validate_ontol_nw}
\alias{validate_ontol_nw}
\title{Validate ontology network for semantic enrichment}
\usage{
validate_ontol_nw(graph)
}
\arguments{
\item{graph}{graph object to validate.}
}
\value{
input graph or validation errors
}
\description{
Performs tests on a graph object in preparation for semantic enrichment.
}
\details{
The tests are:
\enumerate{
   \item Is graph coercible to \code{\link[tidygraph:tidygraph]{tidygraph}} format?
   \item Is graph directed?
   \item Does graph contains one component (is one ontology)?
}
}
