% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eHDPrep-package.R
\docType{package}
\name{eHDPrep-package}
\alias{eHDPrep}
\alias{eHDPrep-package}
\title{'eHDPrep': Quality Control and Semantic Enrichment of Datasets}
\description{
'eHDPrep' is a tool for the preparation and enrichment of health datasets for
analysis. It provides functionality for assessing data quality and for
improving the reliability and machine interpretability of a dataset. 'eHDPrep'
also enables semantic enrichment of a dataset where metavariables are
discovered from the relationships between input variables determined from
user-provided ontologies.
}
\section{Maintainer}{

Ian Overton \email{I.Overton@qub.ac.uk} (\url{https://orcid.org/0000-0003-1158-8527})
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/overton-group/eHDPrep}
  \item Report bugs at \url{https://github.com/overton-group/eHDPrep/issues}
}

}
\author{
Tom Toner \email{ttoner03@qub.ac.uk} (\url{https://orcid.org/0000-0001-8059-5822}),
 Ian Overton \email{I.Overton@qub.ac.uk} (\url{https://orcid.org/0000-0003-1158-8527})
}
\keyword{internal}
